+++ commands/moderation/announce.js
@@ -1,4 +1,8 @@
+'use strict';
+
 const Command = require('../../util/helpers/modules/Command');
+
+//Written by ParadoxOrigins#5451 and Niputi#2490
 
 class Announce extends Command {
     constructor() {
@@ -27,7 +31,7 @@ class Announce extends Command {
                     description: "What's the content of this announcement? You can use the usual markdown, and even masked links using `[masked link](https://google.com)`"
                 },
                 {
-                    description: "Finally, in which channel should i send the announcement?"
+                    description: "Finally, in which channel should I send the announcement?"
                 }
             ]
         };
@@ -44,7 +48,11 @@ class Announce extends Command {
                 color: 0x000,
                 timestamp: new Date()
             };
+            
             embedObject.title = args[0].substr(0, 256);
+            if (!args[1]) {
+                return message.channel.createMessage(':x: You did not specify the color this announcement should take');
+            }
             if (args[1].trim() === "red") {
                 embedObject.color = 0xff0000;
             }
@@ -58,16 +66,21 @@ class Announce extends Command {
                 embedObject.color = parseInt(`0x${args[1].split("#")[1]}`);
                 embedObject.color = embedObject.color === NaN ? 0x000 : embedObject.color;
             }
-
+            else { 
+                embedObject.color = parseInt(`0x${args[1].trim().substr(0,7)}`);
+            }
             embedObject.description = args[2];
-            if (!embedObject.description && !embedObject.title) {
-                return message.channel.createMessage(':x: At least either the title or the description is mandatory');
+            if (!embedObject.description) {
+                return message.channel.createMessage(':x: You did not specify the description this announcement should have');
+            }
+            if (!args[3]) {
+                return message.channel.createMessage(':x: You did not specify where I should send this announcement');
             }
             const channel = await this.getChannelFromText({client: client, message: message, text: args[3]});
             if (!channel) {
                return message.channel.createMessage(':x: I couldn\'t find the channel you specified :v');
             }
-            channel.createMessage({
+            return channel.createMessage({
                 embed: embedObject
             });
     }