+++ commands/music/play.js
@@ -34,23 +34,28 @@ class Play extends Command {
         }
         if (!clientMember.voiceState.channelID) {
             if (Array.isArray(this.clientHasPermissions(message, client, ['voiceConnect', 'voiceSpeak'], message.channel.guild.channels.get(member.voiceState.channelID)))) {
-                return message.channel.createMessage(':x: It seems like i lack the permission to connect or to speak in the voice channel you are in :c');
+                return message.channel.createMessage(':x: It seems like I lack the permission to connect or to speak in the voice channel you are in :c');
             }
         }
-        const player = await client.musicManager.getPlayer(message.channel.guild.channels.get(member.voiceState.channelID));
-        const connection = client.musicManager.connections.get(message.channel.guild.id);
+        const connection = await client.musicManager.getPlayer(message.channel.guild.channels.get(member.voiceState.channelID));
         let track;
         if (!args[0]) {
+            await connection.defer;
             if (connection.queue[0]) {
                 track = connection.queue[0];
+                if (connection.player.paused) {
+                    connection.player.setPause(false);
+                } else {
+                    connection.queue.shift();
+                }
             } else {
                 return message.channel.createMessage(':x: You didn\'t specified any songs to play and there is nothing in the queue');
             }
         }
-        let tracks = track ? [] : await client.musicManager.resolveTracks(player.node, args.join(' '));
+        let tracks = track ? [] : await client.musicManager.resolveTracks(connection.player.node, args.join(' '));
         track = track ? track : tracks[0];
         if (!track) {
-            return message.channel.createMessage(`:x: I could not find any song :c, please make sure to:\n- Follow the syntax (check \`${client.commands.get('help').getPrefix(client, guildEntry)}help ${this.help.name}\`)\n- Use HTTPS links, unsecured HTTP links aren't supported\n- If a YouTube video, i can't play it if it is age-restricted\n - If a YouTube video, it might be blocked in the country my servers are`);
+            return message.channel.createMessage(`:x: I could not find any song :c, please make sure to:\n- Follow the syntax (check \`${client.commands.get('help').getPrefix(client, guildEntry)}help ${this.help.name}\`)\n- Use HTTPS links, unsecured HTTP links aren't supported\n- If a YouTube video, I can't play it if it is age-restricted\n - If a YouTube video, it might be blocked in the country my servers are`);
         }
         if (tracks.length > 1) {
             track = await this.selectTrack(client, message, tracks);
@@ -58,29 +63,9 @@ class Play extends Command {
                 return;
             }
         }
-        player.paused ? await player.setPause(false) : await player.play(track.track);
-        connection.nowPlaying = {
-            info: { 
-                ...track.info,
-                startedAt: Date.now(),
-                requestedBy: message.author.id
-            },
-            track: track.track
-        };
-        return message.channel.createMessage({embed: {
-            title: ':musical_note: Now playing',
-            description: `[${track.info.title}](${track.info.uri})`,
-            fields: [{
-                name: 'Author',
-                value: track.info.author,
-                inline: true
-            }, {
-                name: 'Duration',
-                value: client.musicManager.parseDuration(track),
-                inline: true
-            }],
-            color: client.config.options.embedColor
-        }});
+        connection.play(track, message.author.id);
+        const output = await client.musicManager.genericEmbed(track, connection, 'Now playing');
+        return message.channel.createMessage({embed: output});
     }
 
     async selectTrack(client, message, tracks) {